back/requirements.txt
numpy==1.17.1
SQLAlchemy==1.3.20
Flask==1.1.2
emoji==0.6.0
Flask_Cors==3.0.9
wordcloud==1.8.0
pyodbc==4.0.30
pandas==1.1.5
stanza==1.1.1
pytz==2019.2
tweepy==3.9.0
scikit_image==0.17.2
matplotlib==3.1.1
nltk==3.5
info==0.0.1
Pillow==8.0.1
scikit_learn==0.23.2
